+++ src/main/java/com/example/pdponline/payload/PromoCodeDTO.java
@@ -5,5 +5,5 @@ import lombok.Builder;
 import java.time.LocalDate;
 
 @Builder
-public record PromoCodeDTO(String promoCode, String description, Integer percentage, LocalDate expiryDate,boolean active) {
+public record PromoCodeDTO(String promoCode, String description, Double percentage, LocalDate expiryDate,boolean active) {
 }
